refactor: Rename user feature atom file from atom.ts to atoms.ts
